Allow overriding config options in EvaporateJs 2.0
In EvaporateJs 2.0, the Evaporate.add() function now takes a second argument, overrideConfig. This allows for overriding the global config settings for each request.

This patch allows for specifying a property on AngularEvaporateUpload called overrideConfig. If specified, this will be passed in to Evaporate.add() as a second argument.

I know Angular Evaporate doesn't officially support EvaporateJs 2.0 yet, but I think this patch is a useful workaround until proper support can be added.